requirements.txt
alembic==1.12.0
Flask==2.3.3
google_auth_oauthlib==1.1.0
google-api-python-client==2.98.0
line-bot-sdk==3.4.0
gunicorn==21.2.0
flask_migrate==4.0.4
oauthlib==3.2.2
sqlalchemy==1.4.39
psycopg2==2.9.7
pymysql==1.1.0
Werkzeug==2.3.7
requests==2.31.0
pytz==2022.7
redis==5.0.0
python-dotenv==1.0.0
APScheduler==3.10.4


